st.{h,c}で用いるハッシュ値の型を64bit環境対応のためintからlongに変更した。

+++ slim/src/mc.c
@@ -69,7 +69,7 @@ void state_free(State *s) {
   LMN_FREE(s);
 }
 
-inline int state_hash(State *s) {
+inline long state_hash(State *s) {
   return s->hash;
 }
